runtime(vim): correctly handle ! attrubute for augroup/doautocmd/doautoall
I noticed that
```
:autocmd! FooBar
:augroup! FooBar
```

looks slightly different to how the ! is drawn. While the `!` for the
autocmd case is shown in the same color as autocmd, for :augroup is is
shown as VimEventList

In addition, the vim syntax seems to accept doautocmd! and doautoall!
(with bang attribute) even so this is not valid.

So let's fix all of those and mark the ! as an error for the the
:doautocmd/doautoall case.

Signed-off-by: Christian Brabandt <cb@256bit.org>