runtime(termdebug): add no_tty option to prevent creation of TTY window
In some circumstances, like debugging an embedded system, the TTY window
is useless and there was no way to prevent its creation.

Add a new "no_tty" option to termdebug_config to let user get rid of TTY
window if they don't want it.

This is based on development by @GitMensch in #10389, adapted to newer
Vim Script.

Signed-off-by: Damien Riegel <damien@riegel.io>