patch 9.1.0541: failing test with Vim configured without channel
Problem:  failing test with Vim configured without channel
Solution: In the test Test_null_values() verify that the 'channel' and
          'job' feature is present in Vim. (Dominique Pellé)

Test `Test_null_values` was failing when vim is configured without
channel because the test uses function `null_channel()` (which
requires the `+channel` feature) and using `null_job()` which
requires the `+job` feature. But was reproducible when vim is
configured with:
```
./configure --with-features=huge --enable-gui=none --disable-channel
```

Only check `null_channel()` when the `+channel` feature is available.
Only check `null_job()` when the `+job` feature is available.

closes: #15177

Signed-off-by: Dominique Pellé <dominique.pelle@gmail.com>
Signed-off-by: Christian Brabandt <cb@256bit.org>