fix: fuzzy completion
Simpler alternative to #15193.

How it works: when using fuzzy completion, the completion leader is cut
down to 1 character (if longer), the rest is stored in a "fuzzy
completion leader" that is inserted later, as if the user typed it, so
that the popup menu shows all matches that start with the leader and
match (fuzzily) the rest of the characters.