if_py_both.h: Reset severity for empty messages
Python 3 interpreter sometimes sends error which is not parsable by
`PyArg_Parse()` in `write_output()`, which ends up with empty string
being sent to `writer()` with `emsg_severe` being set, but `emsg()`,
which resets the severity flag, is not called for empty strings.

It can result in certain cases into putting severity on not severe
messages, because severity stays set after processing an empty string -
it was one of the causes of test suite errors when compiled with Python
3.13 and using stable ABI.

The patch resets the severity flag in `writer()` in case the flag is set
and we would have called `emsg()` if the string was not empty.

Partially resolves #15534