make changes to user settings invalidate the cache
Let's say in a terminal buffer we run :HelpToc but it creates a blank
popup because none of our lines match SHELL_PROMPT. We fix that with

    :let g:helptoc = #{shell_prompt '…'}

and rerun :HelpToc. The blank popup appears again! If we run a new
command or modify the buffer some other way, :HelpToc starts working.

This is because of the cached popup buffer: it wasn't invalidated by the
change to g:helptoc.