if_ruby.c: Use variadic functions
Vim fails to build with -std=c23, which is default in GCC 15 in Fedora 42.
It has more strict data type checks, so it takes functions with `()` as `void`,
and then reports errors when function definition introduces arguments.

Using variadic functions makes Vim to be built - do let me know if you would prefer
a different solution.