`setpos` find window attached to target buffer...
Okay so `buf->b_last_cursor` ain't doing the trick, and reading through
other bits of source, and documentation, the cursor position is really
for real stored on a window.

Which means we get the opportunity to search `curwin->w_prev` and/or
`curwin->w_next` for the first window the target buffer is attached to.

This feels closer to correct, though also if it did work feels
sub-optimal, regardless it still misses the mark when mutating a
cursor's state as it seg-faults when touching another tab's buffer.