ci: Make syntax tests not spam output
Currently syntax tests outputs all the Vim commands to the console,
which is annoying for a local developer but also makes the CI output
impossible to parse. Simply redirect all output to /dev/null similar to
script tests. This can be turned off locally for someone diagnosing
issues but it should not be the default behavior to output all terminal
control sequences to output in CI.