feat: add commentstring '> %s' for ftplugin mail
Problem: The new native commenting functionality is currently not used
when editing mail. One could reasonably expect it to change the "quote"
state of any given line in the mail (i.e. the preceding ">"), which
would be very handy and feel natural when editing mail. Especially since
the current file already uses "setlocal comments+=n:>".

Solution: Add commentstring to `> %s` to be used in files of type mail.