Fix :highlight auto-completing cterm colors to -1 when it's unset
Problem: :highlight auto-complete has a minor bug where an existing
highlight group with a cterm color being unset would result in it being
auto-completed to -1 in cmdline which is invalid.

Solution: Correctly check for whether an int value is set to non-zero
before retrieving the existing value for auto-complete. Also do this for
attr/string values as they previously worked only by accident.