feat: pumwidth support -1 for visible alignment
Currently, the width of the longest complete-item is always used for alignment. This wastes some screen space. Give pumwidth a special value of -1. When pumwidth is -1, use the width of items in the visible area of pum for alignment. Still use 15 as the artificial minimum width alignment. When the items beyond 15 are not in the visible area, the width of pum is 15, and the width is adjusted when the elements beyond 15 appear in the area after scrolling.